32-bit support (#9)

+++ include/tmc/asio/aw_asio.hpp
@@ -8,6 +8,7 @@
 // https://github.com/boostorg/cobalt/blob/develop/include/boost/cobalt/op.hpp
 
 #pragma once
+#include "tmc/detail/compat.hpp"
 #include "tmc/detail/concepts.hpp" // IWYU pragma: keep
 #include "tmc/detail/mixins.hpp"
 #include "tmc/detail/thread_locals.hpp"
@@ -107,7 +108,7 @@ template <IsAwAsio Awaitable> struct awaitable_traits<Awaitable> {
     awaitable.customizer.done_count = DoneCount;
   }
 
-  static void set_flags(self_type& awaitable, uint64_t Flags) {
+  static void set_flags(self_type& awaitable, size_t Flags) {
     awaitable.customizer.flags = Flags;
   }
 };
